+++ app/controllers/UserController.php
@@ -27,7 +27,7 @@
  *
  * Problem: Add more function to tradiccional admin.
  * @author $Author: Manuel Gil. $
- * @version $Revision: 0.0.10 $ $Date: 01/29/2021 $
+ * @version $Revision: 0.1.0 $ $Date: 01/30/2021 $
  * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  */
 
@@ -63,7 +63,12 @@ class UserController extends BaseController
 		global $CFG, $USER;
 
 		// Parsing the users.
-		$users = addslashes(json_encode(get_users(), JSON_HEX_AMP | JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_QUOT));
+		$users = addslashes(
+			json_encode(
+				get_users(),
+				JSON_HEX_AMP | JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_QUOT
+			)
+		);
 
 		$params = array(
 			'COMPANY' => COMPANY,
@@ -146,10 +151,10 @@ class UserController extends BaseController
 
 				// Add the new user into the table.
 				$result .= "<tr>
-							<td>{$user->id}</td>
-							<td>{$username}</td>
-							<td>{$password}</td>
-						</tr>";
+								<td>{$user->id}</td>
+								<td>{$username}</td>
+								<td>{$password}</td>
+							</tr>";
 
 				// Add one user to the count.
 				$successes++;
@@ -255,7 +260,12 @@ class UserController extends BaseController
 		global $CFG, $USER;
 
 		// Parsing the users.
-		$users = addslashes(json_encode(get_users_listing(), JSON_HEX_AMP | JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_QUOT));
+		$users = addslashes(
+			json_encode(
+				get_users_listing(),
+				JSON_HEX_AMP | JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_QUOT
+			)
+		);
 
 		$params = array(
 			'COMPANY' => COMPANY,