+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2021-01-30
+
+### Changed
+
+-   Change some file names.
+-   Change query to moodle roles.
+-   Fix issue in roles when list users by course.
+
+## [Unreleased]
+
 ## [0.0.10] - 2021-01-29
 
 ### Added
@@ -104,7 +114,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Initial major release.._
 
-[unreleased]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.10...HEAD
+[unreleased]: https://github.com/ManuelGil/alternate-admin/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.10...v0.1.0
 [0.0.10]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.7...v0.0.8